tests/wpt/tests/css/css-box/parsing/width-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS basic box model: parsing width with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-box-3/#propdef-width">
<meta name="assert" content="width supports only the grammar '[<length> | <percentage>] | available | min-content | max-content | fit-content | auto'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("width", "complex");
test_invalid_value("width", "none");

test_invalid_value("width", "min-content available");
test_invalid_value("width", "max-content 10px");
test_invalid_value("width", "20% available");

test_invalid_value("width", "-10px");
test_invalid_value("width", "-20%");
test_invalid_value("width", "60");
test_invalid_value("width", "10px 20%");

test_invalid_value("width", "10px border-box");
test_invalid_value("width", "content-box 20%");
</script>
</body>
</html>
